Backup current progress
The main takeaway of this commit is the introduction of BookDetailsViewController.

BookDetailsViewController is the controller responsible for displaying the specifics of a specific book such as its title, description, author, artist, list of chapters, and other related information. Chapters are displayed in a table view with plans to add a "compact" view as well that uses a collection view.

For debugging BookDetailsViewController, BookListViewController (which will be responsible for showing all books), only opens the Kaguya book for right now.