Modify BookWindowConstructors
The index of each item is now used as the representedObject so that the action can just query the index in the original array.

Some special care had to be taken with indexes when constructing the chapter list menu because that is reversed but because the chapter list menu is constructed only one time, then there is no trade off in performance.